Fix android build.
This fixes two isseus. The first is that NSS did not support building for
android from Darwin. Correcting the STANDARDS_CFLAGS and using the right `ar`
and `ranlib` addresses that issue.

The second issue is that the path to the toolchain was hardcoded to a version
that is no removed in recnet NDKs.

This is filed upstream as https://bugzilla.mozilla.org/show_bug.cgi?id=911240